% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{AAcoor2NAcoor}
\alias{AAcoor2NAcoor}
\title{AAcoor2NAcoor}
\usage{
AAcoor2NAcoor(frm, AAc1, AAc2)
}
\arguments{
\item{frm}{Frame}

\item{AAc1}{Coordinate 1}

\item{AAc2}{Coordinate 2}
}
\value{

}
\description{
Convert amino acid coordinates to their coorsponding nucleic coordinates (i.e. from region on an ORF to region on a gene)
Transform alignment coordinates from protein locations to ~locationn on the (coding?) nucleic acid sequence.
}
\examples{
AAcoor2NAcoor
}
\seealso{
Other BString_functions: 
\code{\link{AAcoor2NAcoor_dFAST}()},
\code{\link{AAcoor2NAcoor_df}()},
\code{\link{DF2XString}()},
\code{\link{Trim2Core2}()},
\code{\link{XSDNARedunFilter}()},
\code{\link{XString2DF}()},
\code{\link{XString2KmerDT}()}
}
\concept{BString_functions}
