% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{AAcoor2NAcoor_df}
\alias{AAcoor2NAcoor_df}
\title{AAcoor2NAcoor_df}
\usage{
AAcoor2NAcoor_df(dfdt, frmcol, AAc1col, AAc2col, outdf = T)
}
\arguments{
\item{dfdt}{Input table to work on (data.frame or data.table)}

\item{frmcol}{Column name of the field with the frame information.}

\item{AAc1col}{Column name of the field with the start coordinate information.}

\item{AAc2col}{Column name of the field with the end coordinate information.}

\item{outdf}{Should output be a data.frame?}
}
\value{

}
\description{
Like AAcoor2NAcoor, but takes a data.frame as input
}
\examples{
AAcoor2NAcoor_df
}
\seealso{
Other BString_functions: 
\code{\link{AAcoor2NAcoor_dFAST}()},
\code{\link{AAcoor2NAcoor}()},
\code{\link{DF2XString}()},
\code{\link{Trim2Core2}()},
\code{\link{XSDNARedunFilter}()},
\code{\link{XString2DF}()},
\code{\link{XString2KmerDT}()}
}
\concept{BString_functions}
