% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{AAcoor2NAcoor_dFAST}
\alias{AAcoor2NAcoor_dFAST}
\title{AAcoor2NAcoor_dFAST}
\usage{
AAcoor2NAcoor_dFAST(dfdtrw)
}
\arguments{
\item{dfdt}{Input table to work on (data.frame or data.table)rw}
}
\value{
data.frame
}
\description{
Like AAcoor2NAcoor, but takes a data.frame as input AND assumes specific colnames.
}
\examples{
AAcoor2NAcoor_dFAST
}
\seealso{
Other BString_functions: 
\code{\link{AAcoor2NAcoor_df}()},
\code{\link{AAcoor2NAcoor}()},
\code{\link{DF2XString}()},
\code{\link{Trim2Core2}()},
\code{\link{XSDNARedunFilter}()},
\code{\link{XString2DF}()},
\code{\link{XString2KmerDT}()}
}
\concept{BString_functions}
