% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{AllFalse}
\alias{AllFalse}
\title{AllFalse}
\usage{
AllFalse(x)
}
\arguments{
\item{x}{Input logical array/vector}
}
\value{
Logical
}
\description{
Are all items of a logical vector == False?
}
\examples{
AllFalse
}
\seealso{
Other AnyWhich: 
\code{\link{AllTrue}()},
\code{\link{AnyFalse}()},
\code{\link{wana}()},
\code{\link{whd}()},
\code{\link{which.Not.na}()},
\code{\link{which.duplicated}()},
\code{\link{which.na}()},
\code{\link{wh}()},
\code{\link{wna}()}
}
\concept{AnyWhich}
