% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{AllTrue}
\alias{AllTrue}
\title{AllTrue}
\usage{
AllTrue(x)
}
\arguments{
\item{x}{Input logical array/vector}
}
\value{
Logical
}
\description{
Are all items of a logical vector == TRUE?
}
\examples{
AllTrue
}
\seealso{
Other AnyWhich: 
\code{\link{AllFalse}()},
\code{\link{AnyFalse}()},
\code{\link{wana}()},
\code{\link{whd}()},
\code{\link{which.Not.na}()},
\code{\link{which.duplicated}()},
\code{\link{which.na}()},
\code{\link{wh}()},
\code{\link{wna}()}
}
\concept{AnyWhich}
