% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{AnyFalse}
\alias{AnyFalse}
\title{AnyFalse}
\usage{
AnyFalse(x)
}
\arguments{
\item{x}{Logical vector}
}
\value{
Logical
}
\description{
Does the input contain any False values?
}
\examples{
AnyFalse
}
\seealso{
Other AnyWhich: 
\code{\link{AllFalse}()},
\code{\link{AllTrue}()},
\code{\link{wana}()},
\code{\link{whd}()},
\code{\link{which.Not.na}()},
\code{\link{which.duplicated}()},
\code{\link{which.na}()},
\code{\link{wh}()},
\code{\link{wna}()}
}
\concept{AnyWhich}
