% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{CalcPcoverage}
\alias{CalcPcoverage}
\title{CalcPcoverage}
\usage{
CalcPcoverage(hit_df)
}
\arguments{
\item{hit_df}{Input table (data.frame). Usually tabular search results of profile search e.g. hmmsearch}
}
\value{
data.frame
}
\description{
Calculate profile (subject) coverage, based on alignment coordinates and profile length.
}
\examples{
CalcPcoverage
}
\seealso{
Other Table_functions: 
\code{\link{CalcQcoverage}()},
\code{\link{CompareDFDT}()},
\code{\link{HeaderBreakerCb}()},
\code{\link{HeaderBreaker}()},
\code{\link{MergeXtYf}()},
\code{\link{ReadXlsx}()},
\code{\link{RemoveEmptyStrCols}()},
\code{\link{RemoveEmptyStrRows}()},
\code{\link{RemoveNACols}()},
\code{\link{RemovesSparseols}()},
\code{\link{Rename1Col}()},
\code{\link{SharedCols}()},
\code{\link{Trimifo}()},
\code{\link{Trimito}()},
\code{\link{UnSharedCols}()},
\code{\link{WriteWolfTbl}()},
\code{\link{WriteXlsx}()},
\code{\link{dropcols}()},
\code{\link{droprows}()}
}
\concept{Table_functions}
