% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{CreateCSVFromGBFF}
\alias{CreateCSVFromGBFF}
\title{CreateCSVFromGBFF}
\usage{
CreateCSVFromGBFF(input.gbff)
}
\arguments{
\item{input.gbff}{}
}
\value{

}
\description{
Forked from https://github.com/federiva/Monkeys-Working.
The .sh script must be saved in the working directory
}
\examples{
CreateCSVFromGBFF
}
