% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{DF2XString}
\alias{DF2XString}
\title{DF2XString}
\usage{
DF2XString(dfdt, seqcol = "seq", nmcol = "id")
}
\arguments{
\item{dfdt}{Input table to work on (data.frame or data.table)}

\item{seqcol}{Column name for the sequence.}

\item{nmcol}{Column name for the IDs}
}
\value{
BStringSet
}
\description{
Convert data.frame into a BString object.
}
\examples{
DF2XString
}
\seealso{
Other BString_functions: 
\code{\link{AAcoor2NAcoor_dFAST}()},
\code{\link{AAcoor2NAcoor_df}()},
\code{\link{AAcoor2NAcoor}()},
\code{\link{Trim2Core2}()},
\code{\link{XSDNARedunFilter}()},
\code{\link{XString2DF}()},
\code{\link{XString2KmerDT}()}
}
\concept{BString_functions}
