% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{dropcols}
\alias{dropcols}
\title{`dropcols`}
\usage{
dropcols(df, dcols)
}
\arguments{
\item{df}{Input table (data.frame).}

\item{dcols}{Column to remove.}
}
\value{
data.frame
}
\description{
Replacement function to remove columns.
}
\examples{
`dropcols`
}
\seealso{
Other Table_functions: 
\code{\link{CalcPcoverage}()},
\code{\link{CalcQcoverage}()},
\code{\link{CompareDFDT}()},
\code{\link{HeaderBreakerCb}()},
\code{\link{HeaderBreaker}()},
\code{\link{MergeXtYf}()},
\code{\link{ReadXlsx}()},
\code{\link{RemoveEmptyStrCols}()},
\code{\link{RemoveEmptyStrRows}()},
\code{\link{RemoveNACols}()},
\code{\link{RemovesSparseols}()},
\code{\link{Rename1Col}()},
\code{\link{SharedCols}()},
\code{\link{Trimifo}()},
\code{\link{Trimito}()},
\code{\link{UnSharedCols}()},
\code{\link{WriteWolfTbl}()},
\code{\link{WriteXlsx}()},
\code{\link{droprows}()}
}
\concept{Table_functions}
