% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{DropNode}
\alias{DropNode}
\title{DropNode}
\usage{
DropNode(tree, node, include.self = F)
}
\arguments{
\item{tree}{phylo}

\item{include.self}{}
}
\value{

}
\description{
TreeTools::DropTip doesn't work, and ape's drop.tip require labels (doesn't work with internal nodes)
}
\examples{
DropNode
}
