% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{ExtractFastaSeqsFromGBFF}
\alias{ExtractFastaSeqsFromGBFF}
\title{ExtractFastaSeqsFromGBFF}
\usage{
ExtractFastaSeqsFromGBFF(path.to.folder.with.gbff.files)
}
\arguments{
\item{path.to.folder.with.gbff.files}{}
}
\value{

}
\description{
Forked from https://github.com/federiva/Monkeys-Working
The .sh script must be located in the working directory
}
\examples{
ExtractFastaSeqsFromGBFF
}
