% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{GenericRunner}
\alias{GenericRunner}
\title{GenericRunner}
\usage{
GenericRunner(
  command = "echo",
  param = list(threads = 4, mem = 120),
  RunInTmp = F,
  FlgType = "-",
  AsType = " ",
  ...
)
}
\arguments{
\item{command}{Command to call (if notin path, provide the full path to the binary).}

\item{param}{Flags/parameters to pass to the command.
Formatted as a named list, e.g. "flag" = "value".}

\item{RunInTmp}{Should the command be called from a temporary folder made inside the current working directory?}

\item{FlgType}{How should the flags (from "params") be passed to the command? character, e.g. "-", "--"}

\item{AsType}{How should the values of flags (from "params") be joined with the flags? character, e.g. "=", " "}
}
\value{

}
\description{
A simplified way to call executable / shell commands from within R code.
}
\examples{
GenericRunner
}
