% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{GetKids}
\alias{GetKids}
\title{GetKids}
\usage{
GetKids(tree, node, Return_Labels = F, ...)
}
\arguments{
\item{tree}{phylo}

\item{...}{}
}
\value{

}
\description{
Shortend call to phangorn::Descendants
}
\examples{
GetKids
}
