% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{GetNCBITaxonomyChildNodes}
\alias{GetNCBITaxonomyChildNodes}
\title{GetNCBITaxonomyChildNodes}
\usage{
GetNCBITaxonomyChildNodes(nodes.dmp.path, target.taxid)
}
\arguments{
\item{target.taxid}{}
}
\value{

}
\description{
Forked from https://github.com/federiva/Monkeys-Working
nodes.dmp.path is a STRING which is the path to the nodes.dmp
file downloaded from NCBI:
 ftp://ftp.ncbi.nlm.nih.gov/pub/taxonomy/taxdmp.zip
target.taxid is a  STRING which is the upper NCBI taxonomy
 taxid from which we want to get the child's taxids
The output of the function is a .csv file ("|" delimited)
 saved in the current working directory named child_nodes.csv
The function assumes that the .sh script
 GetNCBITaxonomyChildNodes.sh is located in the
 working directory.
}
\examples{
GetNCBITaxonomyChildNodes
}
