% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{GetUnwantedKids}
\alias{GetUnwantedKids}
\title{GetUnwantedKids}
\usage{
GetUnwantedKids(tree, node, omit.internal = T, asids = F, ...)
}
\arguments{
\item{tree}{phylo}

\item{...}{}
}
\value{

}
\description{
GetUnwantedKids
}
\examples{
GetUnwantedKids
}
