% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{Grange2gbRecord}
\alias{Grange2gbRecord}
\title{Grange2gbRecord}
\usage{
Grange2gbRecord(
  Grng,
  sequnces,
  contig_name = "Tuliv",
  contig_length = 24864,
  source_name = "Tuliv",
  circ = "circular",
  DRNA = "DNA",
  organism = "Tuliv",
  keywords = "Virus",
  datee = "10-JAN-2021",
  type = "PHG",
  workdir = "/home/neri/scratch/X2gbk/",
  outputF = "/home/neri/scratch/X2gbk/testing.gbk",
  inputF = "/media/HDD1/uri/projects/Isra/Tuli/New/TuliV.fasta"
)
}
\arguments{
\item{inputF}{}
}
\value{

}
\description{
Attempt to write out a grange object as an NCBI-like Genebank (".gbk") file.
}
\examples{
Grange2gbRecord
}
