% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{HeaderBreaker}
\alias{HeaderBreaker}
\title{HeaderBreaker}
\usage{
HeaderBreaker(dfdt, sep = ".", clb = "id")
}
\arguments{
\item{dfdt}{Input table to work on (data.frame or data.table)}

\item{sep}{Delimiter to split the column by}

\item{clb}{Column (name) to split}
}
\value{
data.frame
}
\description{
Splits a column into multiple columns/nested table.
}
\examples{
HeaderBreaker
}
\seealso{
Other Table_functions: 
\code{\link{CalcPcoverage}()},
\code{\link{CalcQcoverage}()},
\code{\link{CompareDFDT}()},
\code{\link{HeaderBreakerCb}()},
\code{\link{MergeXtYf}()},
\code{\link{ReadXlsx}()},
\code{\link{RemoveEmptyStrCols}()},
\code{\link{RemoveEmptyStrRows}()},
\code{\link{RemoveNACols}()},
\code{\link{RemovesSparseols}()},
\code{\link{Rename1Col}()},
\code{\link{SharedCols}()},
\code{\link{Trimifo}()},
\code{\link{Trimito}()},
\code{\link{UnSharedCols}()},
\code{\link{WriteWolfTbl}()},
\code{\link{WriteXlsx}()},
\code{\link{dropcols}()},
\code{\link{droprows}()}
}
\concept{Table_functions}
