% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{InstallDependenciesBC}
\alias{InstallDependenciesBC}
\title{InstallDependenciesBC}
\usage{
InstallDependenciesBC()
}
\value{

}
\description{
Install and load packages from BioConducter (or tries to)
}
\examples{
InstallDependenciesBC
}
