% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{InstallDependenciesCran}
\alias{InstallDependenciesCran}
\title{InstallDependenciesCran}
\usage{
InstallDependenciesCran()
}
\value{

}
\description{
Installs and load packages from CRAN (or tries to)
}
\examples{
InstallDependenciesCran()
}
