% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{InstallDependenciesGH}
\alias{InstallDependenciesGH}
\title{InstallDependenciesGH}
\usage{
InstallDependenciesGH()
}
\value{

}
\description{
Installs and load packages from GitHub (or tries to)
}
\examples{
InstallDependenciesGH()
}
