% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{isTip}
\alias{isTip}
\title{isTip}
\usage{
isTip(tree, node, ...)
}
\arguments{
\item{tree}{phylo}

\item{node}{node id}

\item{...}{}
}
\value{

}
\description{
Shortend call to treeio::isTip
}
\examples{
isTip
}
