% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{Kalign3}
\alias{Kalign3}
\title{Kalign3}
\usage{
Kalign3(xs, param = NULL)
}
\arguments{
\item{xs}{StringSet}

\item{param}{Flags?}
}
\value{

}
\description{
Shortend caller for kalign3
}
\examples{
Kalign3
}
