% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{MergeXtYf}
\alias{MergeXtYf}
\title{MergeXtYf}
\usage{
MergeXtYf(Xdf, Ydf, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Shortend call to merge with fixed flags (all.x = T, all.y = F)
}
\examples{
MergeXtYf
}
\seealso{
Other Table_functions: 
\code{\link{CalcPcoverage}()},
\code{\link{CalcQcoverage}()},
\code{\link{CompareDFDT}()},
\code{\link{HeaderBreakerCb}()},
\code{\link{HeaderBreaker}()},
\code{\link{ReadXlsx}()},
\code{\link{RemoveEmptyStrCols}()},
\code{\link{RemoveEmptyStrRows}()},
\code{\link{RemoveNACols}()},
\code{\link{RemovesSparseols}()},
\code{\link{Rename1Col}()},
\code{\link{SharedCols}()},
\code{\link{Trimifo}()},
\code{\link{Trimito}()},
\code{\link{UnSharedCols}()},
\code{\link{WriteWolfTbl}()},
\code{\link{WriteXlsx}()},
\code{\link{dropcols}()},
\code{\link{droprows}()}
}
\concept{Table_functions}
