% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{MMseqsClstsReader}
\alias{MMseqsClstsReader}
\title{MMseqsClstsReader}
\usage{
MMseqsClstsReader(inpt = "resultsDB_clu.tsv", clsnm = "rdrp_id")
}
\arguments{
\item{inpt}{Input path.}

\item{clsnm}{}
}
\value{
data.frame
}
\description{
TODO: Rewrite in a sensical manner and write proper documentation.
}
\examples{
MMseqsClstsReader
}
