% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{NID2NLabel}
\alias{NID2NLabel}
\title{NID2NLabel}
\usage{
NID2NLabel(tree, tip)
}
\arguments{
\item{tree}{phylo}

\item{tip}{}
}
\value{

}
\description{
For a phylo object, convert node ID (including tips) to node.label
}
\examples{
NID2NLabel
}
