% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{NLabel2NID}
\alias{NLabel2NID}
\title{NLabel2NID}
\usage{
NLabel2NID(tree, label)
}
\arguments{
\item{tree}{phylo}

\item{label}{}
}
\value{

}
\description{
Shortend call to tidytree::nodeid
}
\examples{
NLabel2NID
}
