% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{pad}
\alias{pad}
\title{pad}
\usage{
pad(
  str,
  pad = " ",
  width = floor(0.9 * getOption("width")),
  side = c("left", "right", "both"),
  use_length = FALSE
)
}
\arguments{
\item{str}{String to pad (character)}

\item{pad}{String to use as prefix/suffix when padding}

\item{width}{Maxiaml Length (number of characters) of pad copies to add to str.}

\item{side}{On which end to append pad}

\item{use_length}{See stringi::stri_pad}
}
\value{

}
\description{
Short call to stringi::stri_pad.
}
\examples{
pad
}
