% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{PopForna}
\alias{PopForna}
\title{PopForna}
\usage{
PopForna(DBN, pop = F, returl = T)
}
\arguments{
\item{returl}{}
}
\value{

}
\description{
Given a DBN enrty, generates (and opens) a URL for forna (RNA secondary structre visulaizer).
}
\examples{
PopForna
}
