% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{Prodigal_AA2NA}
\alias{Prodigal_AA2NA}
\title{Prodigal_AA2NA}
\usage{
Prodigal_AA2NA(ORFlen, ORFstart_nuc, ORFend_nuc, AA1, AA2, Strand)
}
\arguments{
\item{Strand}{}
}
\value{

}
\description{
Prodigal_AA2NA
}
