% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{QuickLoad}
\alias{QuickLoad}
\title{QuickLoad}
\usage{
QuickLoad(
  THREADS = THREADS,
  SessionCatalogPath = "/media/HDD1/uri/RNA_Vir_MTs/V3/Rdatacatalog/"
)
}
\arguments{
\item{THREADS}{Number of cores.}

\item{SessionCatalogPath}{Path to folder on your machine you want to save the RDataFS to.}
}
\value{

}
\description{
Uses the multithearded fastSave to save the current word session as an ".RDataFS" file to some location.
}
\examples{
QuickLoad
}
\seealso{
\code{\link{QuickSave}}
}
