% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{QuickSave}
\alias{QuickSave}
\title{QuickSave}
\usage{
QuickSave(
  THREADS = THREADS,
  SessionCatalogPath = "/media/HDD1/uri/RNA_Vir_MTs/V3/Rdatacatalog/",
  name = "KvPhylo"
)
}
\arguments{
\item{THREADS}{Number of cores.}

\item{SessionCatalogPath}{Path to folder on your machine to read the RDataFS files from.}

\item{name}{File name of output.}
}
\value{

}
\description{
Uses the multithearded fastSave to load the most recently modified ".RDataFS" session from a predefined location.
}
\examples{
QuickSave
}
\seealso{
\code{\link{QuickSave}}
}
