% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{ReadABC}
\alias{ReadABC}
\title{ReadABC}
\usage{
ReadABC(ABCfile, clsnms = c("reprs", "mems"))
}
\arguments{
\item{ABCfile}{Input path.}

\item{clsnms}{}
}
\value{
data.frame
}
\description{
ReadABC
}
\examples{
ReadABC
}
