% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{ReadMcl}
\alias{ReadMcl}
\title{ReadMcl}
\usage{
ReadMcl(mclfile, col1prefix = "motif")
}
\arguments{
\item{mclfile}{Input path.}

\item{col1prefix}{}
}
\value{
data.frame
}
\description{
ReadMcl
}
\examples{
ReadMcl
}
