% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{ReadXlsx}
\alias{ReadXlsx}
\title{ReadXlsx}
\usage{
ReadXlsx(filepath = "tmpout.xlsx", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Shortend call to readxl::read_xlsx
}
\examples{
ReadXlsx
}
\seealso{
Other Table_functions: 
\code{\link{CalcPcoverage}()},
\code{\link{CalcQcoverage}()},
\code{\link{CompareDFDT}()},
\code{\link{HeaderBreakerCb}()},
\code{\link{HeaderBreaker}()},
\code{\link{MergeXtYf}()},
\code{\link{RemoveEmptyStrCols}()},
\code{\link{RemoveEmptyStrRows}()},
\code{\link{RemoveNACols}()},
\code{\link{RemovesSparseols}()},
\code{\link{Rename1Col}()},
\code{\link{SharedCols}()},
\code{\link{Trimifo}()},
\code{\link{Trimito}()},
\code{\link{UnSharedCols}()},
\code{\link{WriteWolfTbl}()},
\code{\link{WriteXlsx}()},
\code{\link{dropcols}()},
\code{\link{droprows}()}
}
\concept{Table_functions}
