% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{RemoveNACols}
\alias{RemoveNACols}
\title{RemoveNACols}
\usage{
RemoveNACols(dfdt)
}
\arguments{
\item{dfdt}{Input table to work on (data.frame or data.table)}
}
\value{
data.frame
}
\description{
Drops columns that contain only NAs.
}
\examples{
RemoveNACols
}
\seealso{
Other Table_functions: 
\code{\link{CalcPcoverage}()},
\code{\link{CalcQcoverage}()},
\code{\link{CompareDFDT}()},
\code{\link{HeaderBreakerCb}()},
\code{\link{HeaderBreaker}()},
\code{\link{MergeXtYf}()},
\code{\link{ReadXlsx}()},
\code{\link{RemoveEmptyStrCols}()},
\code{\link{RemoveEmptyStrRows}()},
\code{\link{RemovesSparseols}()},
\code{\link{Rename1Col}()},
\code{\link{SharedCols}()},
\code{\link{Trimifo}()},
\code{\link{Trimito}()},
\code{\link{UnSharedCols}()},
\code{\link{WriteWolfTbl}()},
\code{\link{WriteXlsx}()},
\code{\link{dropcols}()},
\code{\link{droprows}()}
}
\concept{Table_functions}
