% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{Rename1Col}
\alias{Rename1Col}
\title{Rename1Col}
\usage{
Rename1Col(dfdt, colnm, newcolnm)
}
\arguments{
\item{dfdt}{Input table to work on (data.frame or data.table)}

\item{colnm}{Existing column name}

\item{newcolnm}{New column name}
}
\value{
data.frame
}
\description{
Change the name of a single column.
}
\examples{
Rename1Col
}
\seealso{
Other Table_functions: 
\code{\link{CalcPcoverage}()},
\code{\link{CalcQcoverage}()},
\code{\link{CompareDFDT}()},
\code{\link{HeaderBreakerCb}()},
\code{\link{HeaderBreaker}()},
\code{\link{MergeXtYf}()},
\code{\link{ReadXlsx}()},
\code{\link{RemoveEmptyStrCols}()},
\code{\link{RemoveEmptyStrRows}()},
\code{\link{RemoveNACols}()},
\code{\link{RemovesSparseols}()},
\code{\link{SharedCols}()},
\code{\link{Trimifo}()},
\code{\link{Trimito}()},
\code{\link{UnSharedCols}()},
\code{\link{WriteWolfTbl}()},
\code{\link{WriteXlsx}()},
\code{\link{dropcols}()},
\code{\link{droprows}()}
}
\concept{Table_functions}
