% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{RepositionClade}
\alias{RepositionClade}
\title{RepositionClade}
\usage{
RepositionClade(tree, clade, where, position = NULL)
}
\arguments{
\item{tree}{phylo}

\item{clade}{Internal node id}

\item{where}{Internal node id (to graft to)}

\item{position}{}
}
\value{
phylo
}
\description{
prune a clade from a phylogentic tree and re-graft it at given position
}
\examples{
RepositionClade
}
