% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{ScreenLastz}
\alias{ScreenLastz}
\title{ScreenLastz}
\usage{
ScreenLastz(reference, querry)
}
\arguments{
\item{querry}{}
}
\value{

}
\description{
Forked from Jerome Ambroise script: https://github.com/JeromeAmbroise/Pathogenomics2/blob/master/R/screenfunction.R
}
\examples{
ScreenLastz
}
