% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{show_me_the_contig}
\alias{show_me_the_contig}
\title{show_me_the_contig}
\usage{
show_me_the_contig(full_name)
}
\arguments{
\item{full_name}{A single string of the format "IMG Taxon ID"_"Scaffold ID"}
}
\value{
URL + opens up the web page.
}
\description{
Open up the contig (scaffold) viewer on IMG/MER site.
}
\examples{
show_me_the_contig
}
