% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{SplitTree2MonoPhyloByTips}
\alias{SplitTree2MonoPhyloByTips}
\title{SplitTree2MonoPhyloByTips}
\usage{
SplitTree2MonoPhyloByTips(tree, node)
}
\arguments{
\item{tree}{phylo}

\item{node}{}
}
\value{

}
\description{
SplitTree2MonoPhyloByTips
}
\examples{
SplitTree2MonoPhyloByTips
}
