% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{TabUnroll}
\alias{TabUnroll}
\title{TabUnroll}
\usage{
TabUnroll(dfdt, sep = ",", colnym = "mems", NewColNym = "values")
}
\arguments{
\item{dfdt}{Input table to work on (data.frame or data.table)}

\item{sep}{Delimiter to break the rolled field's cells with.}

\item{colnym}{Field to unroll}

\item{NewColNym}{A name for the field that will contain the unrolled values}
}
\value{
A data.frame.
}
\description{
Inspired by Yuri Wolf' tab_unroll.pl
"Reads the table, unrolls the specified field into separate lines"
}
\examples{
TabUnroll
}
\seealso{
\code{\link{stringi::stri_pad}}
}
