% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{Tree2Edgetbl}
\alias{Tree2Edgetbl}
\title{Tree2Edgetbl}
\usage{
Tree2Edgetbl(tree)
}
\arguments{
\item{tree}{phylo}
}
\value{
data.frame
}
\description{
tabulates a phylo object.
Consider replacnig with tidytree::as_tibble(tree)
}
\examples{
Tree2Edgetbl
}
