% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{Trim2Core2}
\alias{Trim2Core2}
\title{Trim2Core2}
\usage{
Trim2Core2(
  hits_df = Motif_df,
  faa = FL_longest_members,
  add_genetic_code = F,
  Out_faa = T,
  out_df = F,
  input_was = "new_name",
  subject_was = "profile",
  Expand_projectionX = 0
)
}
\arguments{
\item{hits_df}{Input table (data.frame). Usually tabular search results of profile search e.g. hmmsearch}

\item{faa}{Input XString/BString/DNAStringSet...}

\item{add_genetic_code}{If canonical stops are identified within the core region, should this be noted in the output?}

\item{Out_faa}{Should the output include the trimmed StringSet?}

\item{out_df}{Should the output include the potentially modified hit table?}

\item{input_was}{Column name for the input (usually seq IDs).}

\item{subject_was}{Column name for the profile (usually profile IDs).}

\item{Expand_projectionX}{Should the core region be expanded slightly based to cover the entire *projected* region of the alignment?}
}
\value{

}
\description{
Misc functions for extraction of core domain region based on profile matches.
}
\examples{
Trim2Core2
}
\seealso{
Other BString_functions: 
\code{\link{AAcoor2NAcoor_dFAST}()},
\code{\link{AAcoor2NAcoor_df}()},
\code{\link{AAcoor2NAcoor}()},
\code{\link{DF2XString}()},
\code{\link{XSDNARedunFilter}()},
\code{\link{XString2DF}()},
\code{\link{XString2KmerDT}()}
}
\concept{BString_functions}
