% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{Trimifo}
\alias{Trimifo}
\title{Trimifo}
\usage{
Trimifo(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\value{

}
\description{
Misc functions for extraction of core domain region based on profile matches.
}
\examples{
Trimifo
}
\seealso{
Other Table_functions: 
\code{\link{CalcPcoverage}()},
\code{\link{CalcQcoverage}()},
\code{\link{CompareDFDT}()},
\code{\link{HeaderBreakerCb}()},
\code{\link{HeaderBreaker}()},
\code{\link{MergeXtYf}()},
\code{\link{ReadXlsx}()},
\code{\link{RemoveEmptyStrCols}()},
\code{\link{RemoveEmptyStrRows}()},
\code{\link{RemoveNACols}()},
\code{\link{RemovesSparseols}()},
\code{\link{Rename1Col}()},
\code{\link{SharedCols}()},
\code{\link{Trimito}()},
\code{\link{UnSharedCols}()},
\code{\link{WriteWolfTbl}()},
\code{\link{WriteXlsx}()},
\code{\link{dropcols}()},
\code{\link{droprows}()}
}
\concept{Table_functions}
