% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{UnSharedCols}
\alias{UnSharedCols}
\title{UnSharedCols}
\usage{
UnSharedCols(df1, df2)
}
\arguments{
\item{df1}{Table (data.frame)}

\item{df2}{Table (data.frame)}
}
\value{
Character
}
\description{
Column names that  input tables don't have in common.
}
\examples{
UnSharedCols
}
\seealso{
Other Table_functions: 
\code{\link{CalcPcoverage}()},
\code{\link{CalcQcoverage}()},
\code{\link{CompareDFDT}()},
\code{\link{HeaderBreakerCb}()},
\code{\link{HeaderBreaker}()},
\code{\link{MergeXtYf}()},
\code{\link{ReadXlsx}()},
\code{\link{RemoveEmptyStrCols}()},
\code{\link{RemoveEmptyStrRows}()},
\code{\link{RemoveNACols}()},
\code{\link{RemovesSparseols}()},
\code{\link{Rename1Col}()},
\code{\link{SharedCols}()},
\code{\link{Trimifo}()},
\code{\link{Trimito}()},
\code{\link{WriteWolfTbl}()},
\code{\link{WriteXlsx}()},
\code{\link{dropcols}()},
\code{\link{droprows}()}
}
\concept{Table_functions}
