% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{wh}
\alias{wh}
\title{wh}
\usage{
wh(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
shortend to which
}
\examples{
wh
}
\seealso{
Other AnyWhich: 
\code{\link{AllFalse}()},
\code{\link{AllTrue}()},
\code{\link{AnyFalse}()},
\code{\link{wana}()},
\code{\link{whd}()},
\code{\link{which.Not.na}()},
\code{\link{which.duplicated}()},
\code{\link{which.na}()},
\code{\link{wna}()}
}
\concept{AnyWhich}
