% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{whd}
\alias{whd}
\title{whd}
\usage{
whd(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Short call to which.duplicated
}
\examples{
whd
}
\seealso{
Other AnyWhich: 
\code{\link{AllFalse}()},
\code{\link{AllTrue}()},
\code{\link{AnyFalse}()},
\code{\link{wana}()},
\code{\link{which.Not.na}()},
\code{\link{which.duplicated}()},
\code{\link{which.na}()},
\code{\link{wh}()},
\code{\link{wna}()}
}
\concept{AnyWhich}
