% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{which.duplicated}
\alias{which.duplicated}
\title{which.duplicated}
\usage{
which.duplicated(x, returnValue = F, returnalldups = T)
}
\arguments{
\item{x}{Whatever.}

\item{returnValue}{values of duplicated entries}

\item{returnalldups}{Output all  instances/values of duplicated entries?}
}
\value{
If returnValue is True, output is the values of duplicated entries, else output is the indices (numeric vector)
}
\description{
Indices values of duplicated entries in x
}
\examples{
which.duplicated
}
\seealso{
Other AnyWhich: 
\code{\link{AllFalse}()},
\code{\link{AllTrue}()},
\code{\link{AnyFalse}()},
\code{\link{wana}()},
\code{\link{whd}()},
\code{\link{which.Not.na}()},
\code{\link{which.na}()},
\code{\link{wh}()},
\code{\link{wna}()}
}
\concept{AnyWhich}
