% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{which.Not.na}
\alias{which.Not.na}
\title{which.Not.na}
\usage{
which.Not.na(x)
}
\arguments{
\item{x}{Vector/Array (usually).}
}
\value{
Indices (numeric vector)
}
\description{
Indices of the input that are NOT NAs
}
\examples{
which.Not.na
}
\seealso{
Other AnyWhich: 
\code{\link{AllFalse}()},
\code{\link{AllTrue}()},
\code{\link{AnyFalse}()},
\code{\link{wana}()},
\code{\link{whd}()},
\code{\link{which.duplicated}()},
\code{\link{which.na}()},
\code{\link{wh}()},
\code{\link{wna}()}
}
\concept{AnyWhich}
