% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{wna}
\alias{wna}
\title{wna}
\usage{
wna(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Short call to which.na
}
\examples{
wna
}
\seealso{
Other AnyWhich: 
\code{\link{AllFalse}()},
\code{\link{AllTrue}()},
\code{\link{AnyFalse}()},
\code{\link{wana}()},
\code{\link{whd}()},
\code{\link{which.Not.na}()},
\code{\link{which.duplicated}()},
\code{\link{which.na}()},
\code{\link{wh}()}
}
\concept{AnyWhich}
