% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{WriteWolfTbl}
\alias{WriteWolfTbl}
\title{WriteWolfTbl}
\usage{
WriteWolfTbl(dfdt, filepath, ...)
}
\arguments{
\item{dfdt}{Input table to work on (data.frame or data.table)}

\item{...}{}
}
\value{

}
\description{
Save a tab delimited file with heading but no quotes and no rownames.
}
\examples{
WriteWolfTbl
}
\seealso{
Other Table_functions: 
\code{\link{CalcPcoverage}()},
\code{\link{CalcQcoverage}()},
\code{\link{CompareDFDT}()},
\code{\link{HeaderBreakerCb}()},
\code{\link{HeaderBreaker}()},
\code{\link{MergeXtYf}()},
\code{\link{ReadXlsx}()},
\code{\link{RemoveEmptyStrCols}()},
\code{\link{RemoveEmptyStrRows}()},
\code{\link{RemoveNACols}()},
\code{\link{RemovesSparseols}()},
\code{\link{Rename1Col}()},
\code{\link{SharedCols}()},
\code{\link{Trimifo}()},
\code{\link{Trimito}()},
\code{\link{UnSharedCols}()},
\code{\link{WriteXlsx}()},
\code{\link{dropcols}()},
\code{\link{droprows}()}
}
\concept{Table_functions}
