% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{WriteXlsx}
\alias{WriteXlsx}
\title{WriteXlsx}
\usage{
WriteXlsx(dfdt, filepath = "tmpout.xlsx", ...)
}
\arguments{
\item{dfdt}{Input table to work on (data.frame or data.table)}

\item{...}{}
}
\value{

}
\description{
Shortend call to writexl::write_xlsx
}
\examples{
WriteXlsx
}
\seealso{
Other Table_functions: 
\code{\link{CalcPcoverage}()},
\code{\link{CalcQcoverage}()},
\code{\link{CompareDFDT}()},
\code{\link{HeaderBreakerCb}()},
\code{\link{HeaderBreaker}()},
\code{\link{MergeXtYf}()},
\code{\link{ReadXlsx}()},
\code{\link{RemoveEmptyStrCols}()},
\code{\link{RemoveEmptyStrRows}()},
\code{\link{RemoveNACols}()},
\code{\link{RemovesSparseols}()},
\code{\link{Rename1Col}()},
\code{\link{SharedCols}()},
\code{\link{Trimifo}()},
\code{\link{Trimito}()},
\code{\link{UnSharedCols}()},
\code{\link{WriteWolfTbl}()},
\code{\link{dropcols}()},
\code{\link{droprows}()}
}
\concept{Table_functions}
