% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{XString2DF}
\alias{XString2DF}
\title{XString2DF}
\usage{
XString2DF(
  faa,
  input_was = "new_name",
  trimwhite = F,
  seqcolname = "seq",
  addlength = T
)
}
\arguments{
\item{faa}{input xstring/bstring/dnaStringSet...}

\item{input_was}{Column name for the IDs}

\item{trimwhite}{Should whitespaces be removd from sequence names?}

\item{seqcolname}{Column name for the sequence.}

\item{addlength}{Add a "Length" column with the sequences' width?}
}
\value{
data.frame
}
\description{
Convert BString/like objects into a data.frame
}
\examples{
XString2DF
}
\seealso{
Other BString_functions: 
\code{\link{AAcoor2NAcoor_dFAST}()},
\code{\link{AAcoor2NAcoor_df}()},
\code{\link{AAcoor2NAcoor}()},
\code{\link{DF2XString}()},
\code{\link{Trim2Core2}()},
\code{\link{XSDNARedunFilter}()},
\code{\link{XString2KmerDT}()}
}
\concept{BString_functions}
