% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicf.r
\name{XString2KmerDT}
\alias{XString2KmerDT}
\title{XString2KmerDT}
\usage{
XString2KmerDT(
  infaa = BS_all_mot_sssp,
  abcde = c("H", "E", "C"),
  k = 3,
  out_probablity = F,
  input_was = "name_mot"
)
}
\arguments{
\item{infaa}{input xstring/bstring/dnaStringSet...}

\item{abcde}{Alpha-Beit of input strings}

\item{k}{kmer to use (numeric)}

\item{out_probablity}{Should output be normalized as per-sequence probality?}

\item{input_was}{Column name for the IDs}
}
\value{
data.frame
}
\description{
Compute kmer frequency/count per input StringSet.
TODO: replace with Biostrings::oligonucleotideFrequency
}
\examples{
XString2KmerDT
}
\seealso{
Other BString_functions: 
\code{\link{AAcoor2NAcoor_dFAST}()},
\code{\link{AAcoor2NAcoor_df}()},
\code{\link{AAcoor2NAcoor}()},
\code{\link{DF2XString}()},
\code{\link{Trim2Core2}()},
\code{\link{XSDNARedunFilter}()},
\code{\link{XString2DF}()}
}
\concept{BString_functions}
